# Rewriting internals for better performance