--- a/formula/src/main/java/com/instacart/formula/internal/Event.kt
+++ b/formula/src/main/java/com/instacart/formula/internal/Event.kt
@@ -1,0 +1,5 @@
+package com.instacart.formula.internal
+
+fun interface Event {
+    fun dispatch()
+}