Introducing new formula internals which optimize formula action execution.

When using old internals, Formula has to re-evaluate after each state change starting from the top/parent formula all the way down to the child formula that had the state change. For child formulas not in the direct path, they likely won’t have to re-evaluate as long as parent input matches the last one used.

The new formula internals optimize the scenario when during formula evaluation we start an action that emits a state change. Given we are already within the evaluation loop and haven’t returned an output to the parent, we can re-evaluate inline without having to trigger full chain re-evaluation. This enables cutting a bunch of full re-evaluations. This optimization required us to move action start/stop to the evaluation.

Overall, it seems new internals perform about 27% better (and reduce about 80% full re-evaluations).

**Examples of storefront page load as end metric**
| Old | New |
| ---- | --- |
| storefront - formula summary - 369 - total: 370 ms | storefront - formula summary - 73 - total: 270 ms |
| storefront - formula summary - 361 - total: 374 ms | storefront - formula summary - 75 - total: 282 ms |
| storefront - formula summary - 361 - total: 356 ms | storefront - formula summary - 73 - total: 264 ms |
| storefront - formula summary - 359 - total: 450 ms | storefront - formula summary - 77 - total: 308 ms |
| storefront - formula summary - 367 - total: 378 ms | storefront - formula summary - 79 - total: 276 ms |

On average, new internals performs about 100 ms better with 280 ms / 75 runs versus 385 ms / 363 runs.